feat: 잔고 확인

+++ back/c203/src/main/java/com/ssafy/c203/domain/members/repository/MembersRepository.java
@@ -5,7 +5,7 @@ import com.ssafy.c203.domain.members.entity.WithDrawalStatus;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MembersRepository extends JpaRepository<Members, Integer> {
+public interface MembersRepository extends JpaRepository<Members, Long> {
     Members findByEmail(String email);
 
     //userId를 받아 DB 테이블에서 회원을 조회하는 메소드 작성 -> 우리코드에서 id = email
